Move to RenderTile concept in render callbacks.
The old way used to rely on a pixel buffer set up
for the entire render result, then put results in
that. The intermediate buffer would be the one
given to RhinoCycles.

The new way uses the RenderBuffers contained in the
RenderTile instead. The biggest change is that now
the callbacks are given a float array so data can be
handled in the managed world directly.