Fix crash while initializing OpenCL.
With especially older drivers it can happen that
for instance number of devices returned is zero,
but no error code is given by the OpenCL driver.

This leads to a case where device_ids vector is
resized to 0, and then accessed with at(0). This
leads to a crash.

Intel old driver (8.15.10.2696, 3/19/2012) has been
verified to do this.

Protect the platform_ids vector against similar
potential crash.